Fix error in HT1 and preparation of write the HOC
