+++ src/components/Comment.js
@@ -1,9 +1,6 @@
 import React, {Component} from 'react';
 
 class Comment extends Component {
-  constructor(props) {
-      super(props)
-  }
   render(){
     const comment = this.props
       return (
